Get things to work and put checks in right place
