+++ fairlearn/metrics/_metrics_engine.py
@@ -4,6 +4,7 @@
 import numpy as np
 
 from ._group_metric_result import GroupMetricResult
+from ._input_manipulations import _convert_to_1d_array
 
 _MESSAGE_SIZE_MISMATCH = "Array {0} is not the same size as {1}"
 
@@ -39,9 +40,14 @@ def metric_by_group(metric_function, y_true, y_pred, group_membership, sample_we
     y_a = _convert_to_1d_array(y_true)
     y_p = _convert_to_1d_array(y_pred)
     g_d = _convert_to_1d_array(group_membership)
+
+    _check_array_sizes(y_a, y_p, 'y_true', 'y_pred')
+    _check_array_sizes(y_a, g_d, 'y_true', 'group_membership')
+
     s_w = None
     if sample_weight is not None:
         s_w = _convert_to_1d_array(sample_weight)
+        _check_array_sizes(y_a, s_w, 'y_true', 'sample_weight')
 
     # Evaluate the overall metric with the numpy arrays
     # This ensures consistency in how metric_function is called
@@ -111,16 +117,6 @@ def make_group_metric(metric_function):
 
     return wrapper
 
-
-def _convert_to_1d_array(input):
-    result = np.asarray(input)
-
-    if len(result) > 1:
-        result = np.squeeze(result)
-    else:
-        result = result.reshape(1)
-
-    return result
 
 def _check_array_sizes(a, b, a_name, b_name):
     if len(a) != len(b):
